+++ src/compas_timber/connections/l_butt.py
@@ -1,8 +1,8 @@
 from compas_timber.connections.utilities import beam_ref_side_incidence
+from compas_timber.errors import BeamJoinningError
 from compas_timber.fabrication import JackRafterCut
 from compas_timber.fabrication import Lap
 
-from .joint import BeamJoinningError
 from .joint import Joint
 from .solver import JointTopology