+++ src/compas_timber/connections/l_french_ridge_lap.py
@@ -2,9 +2,9 @@ from compas.tolerance import TOL
 
 from compas_timber.connections.utilities import beam_ref_side_incidence
 from compas_timber.connections.utilities import beam_ref_side_incidence_with_vector
+from compas_timber.errors import BeamJoinningError
 from compas_timber.fabrication import FrenchRidgeLap
 
-from .joint import BeamJoinningError
 from .joint import Joint
 from .solver import JointTopology