+++ src/compas_timber/connections/t_butt.py
@@ -1,7 +1,7 @@
-from compas_timber.connections import BeamJoinningError
 from compas_timber.connections import Joint
 from compas_timber.connections import JointTopology
 from compas_timber.connections.utilities import beam_ref_side_incidence
+from compas_timber.errors import BeamJoinningError
 from compas_timber.fabrication import JackRafterCut
 from compas_timber.fabrication import Lap
 
@@ -44,7 +44,7 @@ class TButtJoint(Joint):
         data["mill_depth"] = self.mill_depth
         return data
 
-    def __init__(self, main_beam=None, cross_beam=None, mill_depth=None, **kwargs):
+    def __init__(self, main_beam=None, cross_beam=None, mill_depth=None, fastener=None, **kwargs):
         super(TButtJoint, self).__init__(**kwargs)
         self.main_beam = main_beam
         self.cross_beam = cross_beam
@@ -52,10 +52,37 @@ class TButtJoint(Joint):
         self.cross_beam_guid = kwargs.get("cross_beam_guid", None) or str(cross_beam.guid)
         self.mill_depth = mill_depth
         self.features = []
+        self.fasteners = []
+        if fastener:
+            if fastener.outline is None:
+                fastener = fastener.copy()  # make a copy to avoid modifying the original fastener
+                fastener.set_default(joint=self)
+            self.base_fastener = fastener
+            if self.base_fastener:
+                self.base_fastener.place_instances(self)
+
+    @property
+    def interactions(self):
+        """Returns interactions between elements used by this joint."""
+        interactions = []
+        interactions.append((self.main_beam, self.cross_beam))
+        for fastener in self.fasteners:
+            for interface in fastener.interfaces:
+                if interface is not None:
+                    interactions.append((interface.element, fastener))
+        return interactions
+
+    @property
+    def beams(self):
+        return [self.main_beam, self.cross_beam]
 
     @property
     def elements(self):
-        return [self.main_beam, self.cross_beam]
+        return self.beams + self.fasteners
+
+    @property
+    def generated_elements(self):
+        return self.fasteners
 
     @property
     def cross_beam_ref_side_index(self):
@@ -140,6 +167,12 @@ class TButtJoint(Joint):
             )
             self.cross_beam.add_features(cross_feature)
             self.features.append(cross_feature)
+
+        # add the features applied by the fastener.interfaces
+        for fastener in self.fasteners:
+            for interface in fastener.interfaces:
+                features = interface.get_features(interface.element)
+                interface.element.add_features(features)
 
     def restore_beams_from_keys(self, model):
         """After de-serialization, restores references to the main and cross beams saved in the model."""