+++ src/compas_timber/connections/t_dovetail.py
@@ -6,8 +6,8 @@ from compas_timber.connections.utilities import point_centerline_towards_joint
 from compas_timber.fabrication import DovetailMortise
 from compas_timber.fabrication import DovetailTenon
 from compas_timber.fabrication import TenonShapeType
+from compas_timber.errors import BeamJoinningError
 
-from .joint import BeamJoinningError
 from .joint import Joint
 from .solver import JointTopology